Cat_or_Dog/pyproject.toml
[tool.poetry]
name = "cat_or_dog"
version = "0.1.0"
description = ""
authors = ["Ken <ken@kenh.fr>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8"
matplotlib = ">=3"
numpy = ">=1.19"
pandas = ">=1.2"
tensorflow = ">=2.4"
scikit-learn = ">=0.2"
scikit-image = ">=0.18"
PyQt5 = ">=5"
Pillow = ">=7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
